Refactor dashboard feature to feature-based structure
- Move dashboard components from src/components/dashboard to src/features/dashboard/components
- Move dashboard pages from src/pages to src/features/dashboard/pages
- Update all import statements to reflect new structure
- Add barrel exports in src/features/dashboard/index.ts
- Follow CLAUDE.md recommended project structure guidelines

This refactoring improves code organization by grouping all dashboard-related code into a single feature module, making it more maintainable and scalable.
