+++ src/features/dashboard/components/EmptyDashboard.tsx
@@ -1,9 +1,9 @@
 import React, { useContext } from "react";
 import { useNavigate } from "react-router-dom";
-import { CampaignContext } from "../../context/CampaignContext";
+import { CampaignContext } from "../../../context/CampaignContext";
 import { AiOutlineArrowRight } from "react-icons/ai";
-import { useAuth } from "../../features/auth/context/AuthContext";
-import handleNavigate from "../../helpers/handleNavigate";
+import { useAuth } from "../../auth/context/AuthContext";
+import handleNavigate from "../../../helpers/handleNavigate";
 
 const EmptyDashboard = () => {
    const navigate = useNavigate();
