+++ src/features/dashboard/index.ts
@@ -1,0 +1,10 @@
+// Dashboard feature exports
+export { default as Dashboard } from './pages/Dashboard';
+
+// Dashboard components
+export { default as CampaignsList } from './components/CampaignsList';
+export { default as CampaignReviewButton } from './components/CampaignReviewButton';
+export { default as CardComponant } from './components/CardComponant';
+export { default as CardStats } from './components/CardStats';
+export { default as EmptyDashboard } from './components/EmptyDashboard';
+export { default as NewCardStats } from './components/NewCardStats';
