+++ src/features/dashboard/pages/Dashboard.tsx
@@ -1,14 +1,14 @@
 import React, { useContext, useEffect, useState } from "react";
-import Sidebar from "../components/shared/Sidebar";
-import { useAuth } from "../features/auth/context/AuthContext";
-import { getServiceURL } from "../helpers/getServiceURL";
-import { BrandContext } from "../context/BrandContext";
+import Sidebar from "../../../components/shared/Sidebar";
+import { useAuth } from "../../auth/context/AuthContext";
+import { getServiceURL } from "../../../helpers/getServiceURL";
+import { BrandContext } from "../../../context/BrandContext";
 import { CircularProgress } from "@mui/material";
 import { RiErrorWarningLine } from "react-icons/ri";
-import CampaignsList from "../components/dashboard/CampaignsList";
+import CampaignsList from "../components/CampaignsList";
 import { useNavigate } from "react-router-dom";
-import handleNavigate from "../helpers/handleNavigate";
-import PerformancePredictionDashboard from "../components/performance/PerformancePredictionDashboard";
+import handleNavigate from "../../../helpers/handleNavigate";
+import PerformancePredictionDashboard from "../../../components/performance/PerformancePredictionDashboard";
 import { FaFlask, FaExternalLinkAlt } from "react-icons/fa";
 
 const Dashboard = () => {
